Struct progress_read::ProgressRead

pub struct ProgressRead<R: Read, T, F: FnMut(&T, usize) -> T> {
    source: R,
    accumulator: T,
    progress: F,
}

A reader that reports incremental progress while reading.

Fields
source: Raccumulator: Tprogress: F
Methods

impl<R: Read, T, F: FnMut(&T, usize) -> T> ProgressRead<R, T, F>

Important traits for ProgressRead<R, T, F>
impl<R: Read, T, F: FnMut(&T, usize) -> T> Read for ProgressRead<R, T, F>

pub fn new(source: R, init: T, progress: F) -> ProgressRead<R, T, F>

Construct a new progress reader with the specified underlying reader, initial value for an accumulator, and progress callback.