Struct progress_read::ProgressRead

pub struct ProgressRead<R: Read, T, F: FnMut(&T, usize) -> T> {
    pub(crate) source: R,
    pub(crate) accumulator: T,
    pub(crate) progress: F,
}

A reader that reports incremental progress while reading.

Fields
source: Raccumulator: Tprogress: F
Implementations

impl<R: Read, T, F: FnMut(&T, usize) -> T> ProgressRead<R, T, F>

pub fn new(source: R, init: T, progress: F) -> ProgressRead<R, T, F>

Construct a new progress reader with the specified underlying reader, initial value for an accumulator, and progress callback.